Ensure first match semantics can be properly encoded in config
If we use a "first rule that matches wins" strategy when matching
rules there needs to be a defined order to the rules in the config
file so that this can be tailored. Since Clojure maps are unordered by
definition we need to use a vector of maps to provide the order.

This is actually not a problem for us as we parse in order and never
create an actual Clojure map with those semantics, but this behaviour
should be replicable in Clojure code.

We also add a way to configure the source root(s).